chore: fix types to make the compiler happy (implicit conversions)

+++ src/debug_menu.h
@@ -3,10 +3,10 @@
 #include "settings.h"
 
 struct Debug_Menu {
-    int width_font   = 12;
-    int width_border = 2;
-    int height_box   = SCREEN_HEIGHT * 3.0f/10.0f;
-    int width_box    = SCREEN_WIDTH  * 3.0f/10.0f;
+    f32 width_font   = 12;
+    f32 width_border = 2;
+    f32 height_box   = SCREEN_HEIGHT * 3.0f/10.0f;
+    f32 width_box    = SCREEN_WIDTH  * 3.0f/10.0f;
     bool show        = false;
     Settings& s      = settings;
 };
